Removed extremely unnesscary GameObject field handling.
This seems to have fixed the seemingly random crashes caused by quitting and changing scenes, it seems to have also fixed the hiccup that always occurs whenever the YCTP is pulled up.

I'm gonna be honest with you, I have no clue how such a stupid decision like this was made in the first place. I did tests on this change. HandleField used to be called almost 800,000 times! Now it's down to 4,000. Oh well, at least it's fixed now.